requirements.txt
numpy
pandas
matplotlib
scikit-learn
Flask
gunicorn
imblearn
xgboost
mysql-connector
requests
joblib
pandas
xgboost
